fixed an issue with schedulemanager in master manager

+++ PetNetApp/PetNetApp/Development/Fundraising/ViewCampaignsPage.xaml.cs
@@ -73,7 +73,7 @@ namespace WpfPresentation.Development.Fundraising
         {
             try
             {
-                _fundraisingCampaigns = _masterManager.FundraisingCampaignManager.RetrieveAllFundraisingCampaignsByShelterId(_masterManager.User.ShelterId);
+                _fundraisingCampaigns = _masterManager.FundraisingCampaignManager.RetrieveAllFundraisingCampaignsByShelterId((int)_masterManager.User.ShelterId);
             }
             catch (ApplicationException ex)
             {
